# chore: use ngx.re.match instead of string match to improve performance