--- a/prometheus.lua
+++ b/prometheus.lua
@@ -56,6 +56,7 @@
 -- increments. Copied from https://github.com/Kong/lua-resty-counter
 local resty_counter_lib = require("prometheus_resty_counter")
 local key_index_lib = require("prometheus_keys")
+local ngx_re_match = ngx.re.match
 
 local Prometheus = {}
 local mt = { __index = Prometheus }
@@ -81,6 +82,8 @@ local DEFAULT_BUCKETS = {0.005, 0.01, 0.02, 0.03, 0.05, 0.075, 0.1, 0.2, 0.3,
 
 -- Prefix for internal shared dictionary items.
 local KEY_INDEX_PREFIX = "__ngx_prom__"
+
+local METRICS_KEY_REGEX = [[(.*[,{]le=")(.*)(".*)]]
 
 -- Accepted range of byte values for tailing bytes of utf8 strings.
 -- This is defined outside of the validate_utf8_string function as a const
@@ -277,15 +280,20 @@ end
 -- Returns:
 --   (string) the formatted key
 local function fix_histogram_bucket_labels(key)
-  local part1, bucket, part2 = key:match('(.*[,{]le=")(.*)(".*)')
-  if part1 == nil then
+  local match, err = ngx_re_match(key, METRICS_KEY_REGEX, "jo")
+  if err then
+    self._log_error("failed to match regex: ", err)
+    return
+  end
+
+  if not match then
     return key
   end
 
-  if bucket == "Inf" then
-    return table.concat({part1, "+Inf", part2})
+  if match[2] == "Inf" then
+    return table.concat({match[1], "+Inf", match[3]})
   else
-    return table.concat({part1, tostring(tonumber(bucket)), part2})
+    return table.concat({match[1], tostring(tonumber(match[2])), match[3]})
   end
 end