@tzssangglass 2022-01-11
## before optimization

### flamegraph

![flame-origin](https://user-images.githubusercontent.com/30819887/148881605-c95ea51b-0a69-4520-aedd-92f551f60df9.png)

### test results

```
  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%   11.18ms
 75.000%   17.41ms
 90.000%   23.66ms
 99.000%   45.57ms
 99.900%   92.22ms
 99.990%  123.71ms
 99.999%  136.83ms
100.000%  144.13ms
```

## after optimization

### flamegraph 

![flame-optimize](https://user-images.githubusercontent.com/30819887/148881618-64cd0df5-a11f-4745-895d-8208bd448833.png)

### test results

```
  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%    4.19ms
 75.000%    7.70ms
 90.000%   13.31ms
 99.000%   29.66ms
 99.900%   59.23ms
 99.990%   88.06ms
 99.999%  129.47ms
100.000%  129.79ms
```

Performance has improved, and the reason for this noticeable delay in requests is that I'm putting a lot of pressure on metrics_collect.

The direction of optimization can be continued

1. use more efficient regex
2. refactor this part of the code, preferably to eliminate the regular.

I'm trying.