@tzssangglass 2022-01-13
I checked some of the `string` related functions 

- `string.gsub`: https://github.com/knyar/nginx-lua-prometheus/blob/f0b277ac92129b19d20cda587a8aab911fdca642/prometheus.lua#L170-L176 cannot be replaced by `ngx.re.gsub`, ref: https://github.com/openresty/lua-nginx-module#special-escaping-sequences
- `string.find`:  only plain string searches (no patterns) can be JIT-compiled by LuaJIT. The code is all strings and does not need to be optimised. ref: http://wiki.luajit.org/NYI
- `string.format`: now it's ok.
- `string.sub`:  cannot be optimised, it is supposed to generate new strings.

Although I have added test cases, I hardly know go, so I can't write complex test cases with go ......