@tzssangglass 2022-01-14
> This is great, thank you. Please let me know if you are done with this, and I'll be happy to get it merged.

Yes, it has been done.
 introduces periodic yielding points by `ngx.sleep(0)` will be submitted in the next PR.